+++ Libraries/LibWeb/CMakeLists.txt
@@ -816,6 +816,7 @@ set(SOURCES
     Streams/WritableStreamOperations.cpp
     SVG/AttributeNames.cpp
     SVG/AttributeParser.cpp
+    SVG/Path.cpp
     SVG/SVGAElement.cpp
     SVG/SVGAnimatedEnumeration.cpp
     SVG/SVGAnimatedLength.cpp
