+++ Libraries/LibWeb/HTML/Path2D.cpp
@@ -11,7 +11,7 @@
 #include <LibWeb/Geometry/DOMMatrix.h>
 #include <LibWeb/HTML/Path2D.h>
 #include <LibWeb/SVG/AttributeParser.h>
-#include <LibWeb/SVG/SVGPathElement.h>
+#include <LibWeb/SVG/Path.h>
 
 namespace Web::HTML {
 
@@ -41,7 +41,7 @@ Path2D::Path2D(JS::Realm& realm, Optional<Variant<GC::Root<Path2D>, String>> con
 
     // 4. Let svgPath be the result of parsing and interpreting path according to SVG 2's rules for path data. [SVG]
     auto path_instructions = SVG::AttributeParser::parse_path_data(path->get<String>());
-    auto svg_path = SVG::path_from_path_instructions(path_instructions);
+    auto svg_path = path_instructions.to_gfx_path();
 
     if (!svg_path.is_empty()) {
         // 5. Let (x, y) be the last point in svgPath.
